+++ packages/angular-workspace/nimble-angular/table-column/mapping/tests/nimble-table-column-mapping.directive.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { NimbleTableModule } from '../../../table/nimble-table.module';
 import { NimbleTableColumnMappingModule } from '../nimble-table-column-mapping.module';
 import { NimbleTableColumnMappingDirective, type TableColumnMapping, TableColumnMappingWidthMode } from '../nimble-table-column-mapping.directive';
-import { TableColumnSortDirection } from '../../nimble-table-column-base.directive';
+import { TableColumnPinLocation, TableColumnSortDirection } from '../../nimble-table-column-base.directive';
 
 describe('NimbleTableColumnMapping', () => {
     describe('module', () => {
@@ -30,6 +30,7 @@ describe('NimbleTableColumnMapping', () => {
                         action-menu-slot="my-slot"
                         action-menu-label="my menu"
                         column-hidden="true"
+                        pin-location="left"
                         fractional-width="2"
                         min-pixel-width="40"
                         sort-direction="${TableColumnSortDirection.ascending}"
@@ -93,6 +94,11 @@ describe('NimbleTableColumnMapping', () => {
             expect(nativeElement.columnHidden).toBe(true);
         });
 
+        it('will use template string value for pin-location', () => {
+            expect(directive.pinLocation).toBe(TableColumnPinLocation.left);
+            expect(nativeElement.pinLocation).toBe(TableColumnPinLocation.left);
+        });
+
         it('will use template string values for sortDirection', () => {
             expect(directive.sortDirection).toBe(TableColumnSortDirection.ascending);
             expect(nativeElement.sortDirection).toBe(TableColumnSortDirection.ascending);
@@ -146,6 +152,7 @@ describe('NimbleTableColumnMapping', () => {
                         [actionMenuSlot]="actionMenuSlot"
                         [actionMenuLabel]="actionMenuLabel"
                         [column-hidden]="columnHidden"
+                        [pin-location]="pinLocation"
                         [fractional-width]="fractionalWidth"
                         [min-pixel-width]="minPixelWidth"
                         [sort-direction]="sortDirection"
@@ -170,6 +177,7 @@ describe('NimbleTableColumnMapping', () => {
             public minPixelWidth: number | null = 40;
             public columnId = 'my-column';
             public columnHidden = true;
+            public pinLocation: TableColumnPinLocation = TableColumnPinLocation.none;
             public sortDirection: TableColumnSortDirection = TableColumnSortDirection.ascending;
             public sortIndex: number | null = 0;
             public sortingDisabled = false;
@@ -257,6 +265,17 @@ describe('NimbleTableColumnMapping', () => {
 
             expect(directive.columnHidden).toBe(false);
             expect(nativeElement.columnHidden).toBe(false);
+        });
+
+        it('can be configured with property binding for pin-location', () => {
+            expect(directive.pinLocation).toBeUndefined();
+            expect(nativeElement.pinLocation).toBeUndefined();
+
+            fixture.componentInstance.pinLocation = TableColumnPinLocation.left;
+            fixture.detectChanges();
+
+            expect(directive.pinLocation).toBe(TableColumnPinLocation.left);
+            expect(nativeElement.pinLocation).toBe(TableColumnPinLocation.left);
         });
 
         it('can be configured with property binding for sortDirection', () => {
@@ -404,6 +423,7 @@ describe('NimbleTableColumnMapping', () => {
                         [attr.action-menu-slot]="actionMenuSlot"
                         [attr.action-menu-label]="actionMenuLabel"
                         [attr.column-hidden]="columnHidden"
+                        [attr.pin-location]="pinLocation"
                         [attr.fractional-width]="fractionalWidth"
                         [attr.min-pixel-width]="minPixelWidth"
                         [attr.sort-direction]="sortDirection"
@@ -428,6 +448,7 @@ describe('NimbleTableColumnMapping', () => {
             public minPixelWidth: number | null = 40;
             public columnId = 'my-column';
             public columnHidden = true;
+            public pinLocation: TableColumnPinLocation = TableColumnPinLocation.none;
             public sortDirection: TableColumnSortDirection = TableColumnSortDirection.ascending;
             public sortIndex: number | null = 0;
             public sortingDisabled = false;
@@ -515,6 +536,17 @@ describe('NimbleTableColumnMapping', () => {
 
             expect(directive.columnHidden).toBe(false);
             expect(nativeElement.columnHidden).toBe(false);
+        });
+
+        it('can be configured with attribute binding for pin-location', () => {
+            expect(directive.pinLocation).toBeUndefined();
+            expect(nativeElement.pinLocation).toBeUndefined();
+
+            fixture.componentInstance.pinLocation = TableColumnPinLocation.left;
+            fixture.detectChanges();
+
+            expect(directive.pinLocation).toBe(TableColumnPinLocation.left);
+            expect(nativeElement.pinLocation).toBe(TableColumnPinLocation.left);
         });
 
         it('can be configured with attribute binding for sortDirection', () => {