+++ packages/nimble-components/src/table/models/table-layout-manager.ts
@@ -1,6 +1,7 @@
 import { observable } from '@ni/fast-element';
 import type { Table } from '..';
 import type { TableColumn } from '../../table-column/base';
+import { TableColumnPinLocation } from '../types';
 import type { TableRecord } from '../types';
 
 /**
@@ -35,7 +36,7 @@ export class TableLayoutManager<TData extends TableRecord> {
     public constructor(private readonly table: Table<TData>) {}
 
     public getGridTemplateColumns(): string {
-        return this.getVisibleColumns()
+        return this.getUnpinnedVisibleColumns()
             .map(column => {
                 const {
                     minPixelWidth,
@@ -77,7 +78,7 @@ export class TableLayoutManager<TData extends TableRecord> {
         this.activeColumnIndex = this.leftColumnIndex + (this.activeColumnDivider % 2);
         this.dragStart = dragStart;
         this.currentTotalDelta = 0;
-        this.visibleColumns = this.getVisibleColumns();
+        this.visibleColumns = this.getUnpinnedVisibleColumns();
         this.setColumnsToFixedSize();
         this.initialTableScrollableWidth = this.table.viewport.scrollWidth;
         this.initialTableScrollableMinWidth = this.table.tableScrollableMinWidth;
@@ -203,7 +204,7 @@ export class TableLayoutManager<TData extends TableRecord> {
      */
     private getFirstLeftResizableColumnIndex(columnIndex: number): number {
         const visibleColumns = this.visibleColumns.length === 0
-            ? this.getVisibleColumns()
+            ? this.getUnpinnedVisibleColumns()
             : this.visibleColumns;
         for (let i = columnIndex; i >= 0; i--) {
             const column = visibleColumns[i];
@@ -224,7 +225,7 @@ export class TableLayoutManager<TData extends TableRecord> {
      */
     private getFirstRightResizableColumnIndex(columnIndex: number): number {
         const visibleColumns = this.visibleColumns.length === 0
-            ? this.getVisibleColumns()
+            ? this.getUnpinnedVisibleColumns()
             : this.visibleColumns;
         for (let i = columnIndex; i < visibleColumns.length; i++) {
             const column = visibleColumns[i];
@@ -361,8 +362,11 @@ export class TableLayoutManager<TData extends TableRecord> {
         }
     }
 
-    private getVisibleColumns(): TableColumn[] {
-        return this.table.columns.filter(column => !column.columnHidden);
+    private getUnpinnedVisibleColumns(): TableColumn[] {
+        return this.table.columns.filter(
+            column => !column.columnHidden
+                && column.columnInternals.pinLocation !== TableColumnPinLocation.left
+        );
     }
 
     private getLeftColumnIndexFromDivider(dividerIndex: number): number {