+++ packages/nimble-components/src/table/models/table-validator.ts
@@ -1,5 +1,6 @@
 import type { TableColumn } from '../../table-column/base';
 import {
+    TableColumnPinLocation,
     type TableRecord,
     TableRowSelectionMode,
     type TableSetRecordHierarchyOptions,
@@ -21,6 +22,7 @@ export class TableValidator<TData extends TableRecord> {
     private idFieldNameNotConfigured = false;
     private invalidColumnConfiguration = false;
     private invalidParentIdConfiguration = false;
+    private invalidPinnedColumnConfiguration = false;
 
     private readonly recordIds = new Set<string>();
 
@@ -35,7 +37,8 @@ export class TableValidator<TData extends TableRecord> {
             duplicateGroupIndex: this.duplicateGroupIndex,
             idFieldNameNotConfigured: this.idFieldNameNotConfigured,
             invalidColumnConfiguration: this.invalidColumnConfiguration,
-            invalidParentIdConfiguration: this.invalidParentIdConfiguration
+            invalidParentIdConfiguration: this.invalidParentIdConfiguration,
+            invalidPinnedColumnConfiguration: this.invalidPinnedColumnConfiguration
         };
     }
 
@@ -153,6 +156,17 @@ export class TableValidator<TData extends TableRecord> {
             x => !x.columnInternals.validator.isColumnValid
         );
         return !this.invalidColumnConfiguration;
+    }
+
+    public validatePinnedColumnConfigurations(
+        columns: readonly TableColumn[]
+    ): boolean {
+        this.invalidPinnedColumnConfiguration = columns.some(
+            x => x.columnInternals.pinLocation === TableColumnPinLocation.left
+                && (x.columnInternals.pixelWidth === undefined
+                    || !x.columnInternals.resizingDisabled)
+        );
+        return !this.invalidPinnedColumnConfiguration;
     }
 
     public getPresentRecordIds(